Apply the `kotlin-dsl` plugin to `buildSrc`
The `kotlin-dsl` plugin

- applies the `org.jetbrains.kotlin.jvm` plugin,
- configures an embedded repository that contains all Kotlin libraries embedded into Gradle,
- and pins them to the embedded Kotlin version,
- it also adds compile only dependencies on `kotlin-stdlib` and `kotlin-reflect`,
- and adds the `gradleKotlinDsl()` dependency to the `compileOnly` and `testRuntimeOnly` configurations,
- and configures the Kotlin DSL compiler plugins

This commit also includes changes to:
- gradle/compile.gradle
- gradle/strictCompile.gradle

Those scripts assumed all `AbstractCompile` tasks to have an `options`
property. This is not the case. Java, Groovy and Scala compilation tasks
do, but the KotlinCompile tasks now present in `buildSrc` don't.

gradle/compile.gradle now target only the tasks we are interested by, that is Java and Groovy compilation tasks.
There are no scala compilation tasks in the Gradle build itself.
The TaskCollection is set as an extra property reused by gradle/strictCompile.gradle.

Signed-off-by: Paul Merlin <paul@gradle.com>